all: swap out the C++ ethash to the pure Go one (mining todo)
