+++ log/format.go
@@ -23,7 +23,6 @@ const (
 // locationTrims are trimmed for display to avoid unwieldy log lines.
 var locationTrims = []string{
 	"github.com/ethereum/go-ethereum/",
-	"github.com/ethereum/ethash/",
 }
 
 // PrintOrigins sets or unsets log location (file:line) printing for terminal
